Add files via upload
In this thesis, our primary focus was to study the PageRank algorithm, its formulation, 
justification and other different variations of it. In our early days we studied about this 
algorithm and its variations. We also justified the formulation of the algorithm by 
implementing the PageRank on a small sample web structure.
In the beginning we started our thesis by learning the very basic PageRank algorithm 
developed by Larry Page and Sergey Brin. Because it was the first ever formulation that 
was used to find out the PageRank of websites. We have implemented this basic idea by 
using a small web structure and also analyzed its features, properties, strengths and 
weaknesses.
After gathering well enough knowledge we explored some other variations of the 
PageRank algorithm. We also analyzed the properties thoroughly.
Then we moved into the topic Search Engine Optimization (SEO). We studied different 
criteria of the SEO technique. We then implemented these criteria on two experimental 
websites and found out the results of them.
After that, we worked on a new technique for SEO. We used Word2Vec which is one 
kind of topic modeling in case of Natural Language Processing to find out probable 
keywords to use on our experimental website to differentiate between them. We 
successfully implemented it.